Restructure folder layout of home screen
