+++ test/full_coverage_test.dart
@@ -100,14 +100,15 @@ import 'package:pi_hole_client/screens/domains/widgets/add_domain_modal.dart';
 import 'package:pi_hole_client/screens/domains/widgets/domain_details_screen.dart';
 import 'package:pi_hole_client/screens/domains/widgets/domain_tile.dart';
 import 'package:pi_hole_client/screens/domains/widgets/domains_list.dart';
-import 'package:pi_hole_client/screens/home/disable_modal.dart';
+import 'package:pi_hole_client/screens/domains/widgets/edit_domain_modal.dart';
 import 'package:pi_hole_client/screens/home/home.dart';
-import 'package:pi_hole_client/screens/home/home_appbar.dart';
-import 'package:pi_hole_client/screens/home/home_charts.dart';
-import 'package:pi_hole_client/screens/home/home_tile.dart';
-import 'package:pi_hole_client/screens/home/switch_server_modal.dart';
 import 'package:pi_hole_client/screens/home/widgets/clients_last_hours_line.dart';
+import 'package:pi_hole_client/screens/home/widgets/disable_modal.dart';
+import 'package:pi_hole_client/screens/home/widgets/home_appbar.dart';
+import 'package:pi_hole_client/screens/home/widgets/home_charts.dart';
+import 'package:pi_hole_client/screens/home/widgets/home_tile.dart';
 import 'package:pi_hole_client/screens/home/widgets/queries_last_hours_line.dart';
+import 'package:pi_hole_client/screens/home/widgets/switch_server_modal.dart';
 import 'package:pi_hole_client/screens/logs/clients_filters_modal.dart';
 import 'package:pi_hole_client/screens/logs/log_details_screen.dart';
 import 'package:pi_hole_client/screens/logs/log_status.dart';
